requirements.txt
huggingface_hub==0.32.3
numpy==2.3.0
opencv_python==4.11.0.86
pandas==2.3.0
PyYAML==6.0.2
PyYAML==6.0.2
skimage==0.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
